requirements.txt
Flask
pandas
scikit-learn
requests
beautifulsoup4
python-dotenv